Add initial Rollup config/scripts
